+++ packages/cdn/public/forms/cdn_region-1-fm.vue
@@ -1,9 +1,11 @@
 <template>
-  <div class="u-form-layout">
+  <div
+    v-loading="loading"
+    class="u-form-layout"
+  >
     <u-toolbar />
 
     <u-form-container
-      v-loading.body="loading"
       label-position="top"
       :max-width="800"
     >